hostip: Fix CURLOPT_RESOLVE to append to an existing address list
(Curl_loadhostpairs)
- If a host:port cache entry already exists append to its address list
instead of discarding the new address.

Prior to this change only the first address passed to CURLOPT_RESOLVE
for each host:port cache entry was added to its address list.

    host = curl_slist_append(NULL, "google.com:80:167.206.145.177");
    host = curl_slist_append(host, "google.com:80:167.206.145.178");
    curl_easy_setopt(curl, CURLOPT_RESOLVE, host);

In that case only 167.206.145.177 would have been added to the cache
entry for google.com:80. This commit changes the behavior so that both
would be added.