schannel: alternate solution for issue 244. draft7
- Regardless of len value go directly to cleanup if there is an
unrecoverable error or a close_notify was already received. Prior to
this change we only acknowledged those two states if len != 0.

- Fix a bug in connection closed behavior: Set the error state in the
cleanup, because we don't know for sure it's an error until that time.

- (Related to above) In the case the connection is closed go "greedy"
with the decryption to make sure all remaining encrypted data has been
decrypted even if it is not needed at that time by the caller. This is
necessary because we can only tell if the connection closed gracefully
(close_notify) once all encrypted data has been decrypted.

- Do not renegotiate when an unrecoverable error is pending.

- Fix a bug from draft 6 in the renegotiation behavior: Change the
sspi_status back to SEC_E_OK on renegotiation success otherwise the loop
will not continue.

- Fix a bug from draft6 in setting the unrecoverable_err state: The
unrecoverable error state must be set before any return, otherwise it
could be lost.