openssl: Integrate Peter Wu's SSLKEYLOGFILE implementation. draft1
I adapted Peter Wu's SSLKEYLOGFILE implementation for libcurl's
openssl.c so that I can dump ssl master secrets to a logfile.

*His code is GPL licensed, do not merge in master.*

For this to work set environment variable SSLKEYLOGFILE to a filename.
And in Wireshark: Edit > Preferences > Protocols > SSL > Master-Secret

https://git.lekensteyn.nl/peter/wireshark-notes/tree/src/sslkeylog.c

This commit is mostly to help diagnose an issue we are working on in
libcurl, 'HTTP/2 final frames not processed on server disconnect'. It
also changes http2.c on_frame_recv to dump the frame name in debug
builds of libcurl.