curl: disable output files archive bit until close (Windows)
(***Draft***)
Doesn't work fully. Windows still turns on archive bit even if
FILE_ATTRIBUTE_NORMAL is used to create the file, regardless of whether
or not it already exists. Also even if the archive bit is removed after
creation Windows can turn it back on whenever the file's buffer is
written to disk, ie modified.

- Change the output files created by the curl tool via
  tool_create_output_file() to disable the file's archive bit before
  writing to the file and enable it before closing the file.

The curl tool on Windows opens output files in a way that other
processes can read the file as it's being written, ie fopen default
shared read. Prior to this change a backup process that depends on the
archive bit could see fit to backup an output file before we are
finished writing it. Therefore we now disable the file's archive bit
before writing to it and enable it before closing the file.

Fixes https://github.com/curl/curl/issues/3354
Closes #xxxx