http2: hack to fix http proxy over http2. draft1
- Parse the scheme, host and path from the :path component and then
  use that information to update the respective pseudo-headers.

http2.c constructs an http2 request by parsing a regular http request
that has been constructed in http.c. Normally this request has the
path as the request-uri field (ie /foo in GET /foo HTTP/1.1) so that is
what is used for http2 pseudo-header :path. However in the case of an
http proxy libcurl uses the full URL as the request-uri field
(ie http://test.com/foo in GET http://test.com/foo HTTP/1.1) so that
is not just :path but must be split into :scheme and :authority (host)
which is what this patch does. A side effect of this is that any
user-specified host field is overridden.

Ref: https://github.com/curl/curl/issues/3570#issuecomment-464304328
Ref: https://curl.haxx.se/docs/todo.html#Support_HTTP_2_for_HTTP_S_proxi