build: support Windows CRT heap memory tracking
- Support Windows CRT heap memory tracking (_CRTDBG_MAP_ALLOC defined)
  as an alternative to libcurl heap memory tracking (CURLDEBUG defined).

This commit is to demonstrate how it's possible to use Windows CRT heap
memory tracking to detect leaks in libcurl *and* the application, with
filenames and line numbers recorded for the leaks.

This is not intended to go upstream, it was written to further the
discussion.

See winbuild\README_HEAP_DEBUG.md for instructions.

Ref: https://github.com/curl/curl/issues/12327
Ref: https://github.com/jay/curl/blob/crtdbg/winbuild/README_HEAP_DEBUG.md